refactor: Unify restorable swap map
