+++ src/context/Rescue.tsx
@@ -2,16 +2,16 @@ import type { ECPairInterface } from "ecpair";
 import type { Accessor, JSX, Setter } from "solid-js";
 import { createContext, createSignal, useContext } from "solid-js";
 
-import type { RestorableSwap } from "../utils/boltzClient";
 import { ECPair } from "../utils/ecpair";
 import { type RescueFile, deriveKey } from "../utils/rescueFile";
+import type { SomeSwap } from "../utils/swapCreator";
 
 export type RescueContextType = {
     rescueFile: Accessor<RescueFile>;
     setRescueFile: Setter<RescueFile>;
 
-    rescuableSwaps: Accessor<RestorableSwap[]>;
-    setRescuableSwaps: Setter<RestorableSwap[]>;
+    rescuableSwaps: Accessor<SomeSwap[]>;
+    setRescuableSwaps: Setter<SomeSwap[]>;
 
     deriveKey: (index: number) => ECPairInterface;
 };
@@ -20,9 +20,7 @@ const RescueContext = createContext<RescueContextType>();
 
 export const RescueProvider = (props: { children: JSX.Element }) => {
     const [rescueFile, setRescueFile] = createSignal<RescueFile>();
-    const [rescuableSwaps, setRescuableSwaps] = createSignal<RestorableSwap[]>(
-        [],
-    );
+    const [rescuableSwaps, setRescuableSwaps] = createSignal<SomeSwap[]>([]);
 
     const deriveKeyWrapper = (index: number) => {
         return ECPair.fromPrivateKey(
