+++ src/pages/RefundRescue.tsx
@@ -1,72 +1,18 @@
 import { useParams } from "@solidjs/router";
-import { OutputType } from "boltz-core";
 import log from "loglevel";
 import type { Accessor } from "solid-js";
 import { Show, createResource, createSignal, onCleanup } from "solid-js";
 
 import BlockExplorer from "../components/BlockExplorer";
 import RefundButton from "../components/RefundButton";
-import { SwapType } from "../consts/Enums";
 import { useGlobalContext } from "../context/Global";
 import { usePayContext } from "../context/Pay";
 import { useRescueContext } from "../context/Rescue";
-import type { ChainSwapDetails, RestorableSwap } from "../utils/boltzClient";
 import { getSwapStatus } from "../utils/boltzClient";
 import { ECPair } from "../utils/ecpair";
 import { getRefundableUTXOs } from "../utils/rescue";
 import { deriveKey } from "../utils/rescueFile";
 import type { ChainSwap, SomeSwap, SubmarineSwap } from "../utils/swapCreator";
-
-export const mapSwap = (
-    swap?: RestorableSwap,
-): Partial<SomeSwap> | undefined => {
-    if (swap === undefined) {
-        return undefined;
-    }
-
-    switch (swap.type) {
-        case SwapType.Submarine:
-            return {
-                ...swap,
-                swapTree: swap.refundDetails.tree,
-                assetSend: swap.from,
-                assetReceive: swap.to,
-                version: OutputType.Taproot,
-                blindingKey: swap.refundDetails.blindingKey,
-                address: swap.refundDetails.lockupAddress,
-                refundPrivateKeyIndex: swap.refundDetails.keyIndex,
-                claimPublicKey: swap.refundDetails.serverPublicKey,
-            };
-        case SwapType.Chain:
-            return {
-                ...swap,
-                assetSend: swap.from,
-                assetReceive: swap.to,
-                version: OutputType.Taproot,
-                address: swap.claimDetails.lockupAddress,
-                refundPrivateKeyIndex: swap.refundDetails.keyIndex,
-                claimPublicKey: swap.claimDetails.serverPublicKey,
-                claimPrivateKeyIndex: swap.claimDetails.keyIndex,
-                lockupDetails: {
-                    ...swap.refundDetails,
-                    swapTree: swap.refundDetails.tree,
-                } as ChainSwapDetails,
-            };
-        case SwapType.Reverse:
-            return {
-                ...swap,
-                assetSend: swap.from,
-                assetReceive: swap.to,
-                version: OutputType.Taproot,
-                address: swap.claimDetails.lockupAddress,
-                claimPublicKey: swap.claimDetails.serverPublicKey,
-                claimPrivateKeyIndex: swap.claimDetails.keyIndex,
-                sendAmount: swap.claimDetails.amount,
-            };
-        default:
-            return undefined;
-    }
-};
 
 const RefundRescue = () => {
     const params = useParams<{ id: string }>();
@@ -83,7 +29,7 @@ const RefundRescue = () => {
     const { rescuableSwaps, rescueFile } = useRescueContext();
 
     const rescuableSwap = () =>
-        mapSwap(rescuableSwaps().find((swap) => swap.id === params.id));
+        rescuableSwaps().find((swap) => swap.id === params.id);
 
     const [refundTxId, setRefundTxId] = createSignal<string>("");
